requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Admin==1.6.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
yadisk==2.1.0
yookassa==3.0.0
python-dotenv==1.0.1
email-validator==2.1.0.post1
Werkzeug==3.0.1
SQLAlchemy==2.0.28
alembic==1.13.1
python-slugify==8.0.4
gunicorn==21.2.0
mod_wsgi==5.0.0 